Update Parser and AddCommand to handle new format
As discussed in the updated UserGuide, uMessage will accept new contacts
in a different format from that of AB3.

In particular, add command no longer has a need for `a/`, `p/` and `e/`
prefixes. Additionally, the add command will accept a variable number of
`d/` prefixes, which specifies a social media account/communication
channel that the person has.