Fix SocialMedia models not being stored in JSON
Previously, test cases passed because Person did not set its socialMedia
attribute in the constructor. Upon fixing said oversight, it appears
that SocialMedia was not properly serialized into JSON format.

This commit ensures that now SocialMedia models can be serialized, and
also fixes a few test cases that will fail because of the new serialized
format.