Refactor IndexedDeleteCommand,  NamedDeleteCommand
Since we will be implementing several commands that can target the list in AddressBook based on an Index or Name of Person, we should
refactor this targetting into a separate class.

This reduces the need to make two sets of commands whenever we wish to allow for targetting by
both Index and Name.