Add two new observable values in Model
currentlySelectedPersonProperty represents the currently selected person
in the addressbook. The currently selected person is the one displayed in
the RHS panel at any given moment. The RHS panel will observe this
property for any changes, and update accordingly if the person
encapsulated by this variable changes.

selectionIndex represents the current index selected in PersonListPanel.
Typically, the ListView in PersonListPanel already exposes a
SelectionModel for us to track and modify the current selection in a
ListView. But since we want to reduce coupling and ensure that Ui only
observes and reacts to changes in the model, we should not track the
SelectionModel in ListView, but track a separate observable value.
PersonListPanel will observe this variable, and update accordingly if
the selectionIndex is changed.