Fix Edit command
Edit replaces the old Person instance with a new Person instance, but
the currentlySelectedPerson still references the old Person instance.

This causes a visual bug where if the currentlySelectedPerson
is the one being edited, the RHS panel does not update to reflect the
new edits to the person.

Furthermore, we should show the newly edited person instance to the user
so that they can be assured of the changes. We also set the
selectionIndex to null so the LHS does not have the selection bar over
the wrong Person.

This commit also updates ViewCommand to work with previous changes to
setSelectedIndex.