Fix css of ListCell and ListView
Fixes #141.

Change so that the VBox of the PersonListCard does not dictate the
background color of the cardpane anymore. Instead, rely on the .list-cell
and .list-view in LightTheme.css to set the background color of the
cell.

Also, change javafx17 to javafx11 to prevent incompatibility warnings
from happening since the max supported javafx version on some computers is
lower than 17.