Fix target type confusion in Target.java
Previously, a integer string representing a valid name would be
interpreted as an index by the parser, causing the user to not be able
to refer to a Person who has a number as their name by their name.

Likewise, a valid mathematical integer but an invalid index would be
interpreted as a name rather than an index. Examples include negative
integers and integers beyond 2^31 - 1.

Now, the Target class will first interpret the given string first as an
Index, failing which will cause it to treat it as a Name, and failing
which again will cause the Target class to throw a ParseException
because of an invalid name, unless the string is a mathematical integer,
which would throw a ParseException due to invalid index.

After ensuring that the given string is a valid Name (all valid indexes
are valid names), the Target class will first interpret the given string
as a Name to see if there are any matches in the list with the same
Name.

If there are matches and the string is also a valid index within the
bounds of the list, then the string is treated as an index, otherwise a
name.

If there are no matches by name and the string is an integer, treat the
string as an index, and throw index exceptions accordingly.

If there are matches by name and the string is not an integer, treat the
string as a name, and throw exceptions due to name accordingly.