Handle edge case in Target class
Previously, some non-negative integers (0, MAX_INT to INFINITY) are
invalid indices, but are valid names. Thus, they will not be caught as
invalid index during the instantiation of the Target class, since they
could very well represent a valid name at runtime when the target list
is created.

There is no choice but to defer the validation to runtime. If these
values do not represent a name in the list at runtime, then they need to
be caught and treated as invalid indices at runtime.