Remove invalid test from EditCommandParser
As mentioned in commit 84c21baf693, 0 will not be caught as an invalid
index by the parser at parsing time.

"0" will only be caught as an invalid index when there is no person in the
list with the name 0, and can only be caught at runtime. Thus,
checking whether 0 (or any integer bigger than 2^31) is a valid index in
the parser is invalid.